Cleanup the testing infra for gradle plugin
