+++ .github/workflows/main.yml
@@ -73,3 +73,12 @@ jobs:
       with:
         name: integration-tests-reports
         path: integration-tests/build/reports
+    # Run gradle plugin tests
+    - name: gradle plugin tests
+      run: ./gradlew --info :gradle-plugin:test
+    - name: Upload gradle plugin test results
+      if: always()
+      uses: actions/upload-artifact@v2
+      with:
+        name: gradle-plugin-test-reports
+        path: gradle-plugin/build/reports
