+++ integration-tests/src/test/kotlin/com/google/devtools/ksp/test/PlaygroundIT.kt
@@ -17,7 +17,7 @@ class PlaygroundIT {
     fun testPlayground() {
         val gradleRunner = GradleRunner.create().withProjectDir(project.root)
 
-        val resultCleanBuild = gradleRunner.withArguments("clean", "build").build()
+        val resultCleanBuild = gradleRunner.withArguments("clean", "build", "--stacktrace").build()
 
         Assert.assertEquals(TaskOutcome.SUCCESS, resultCleanBuild.task(":workload:build")?.outcome)
