+++ symbol-processing/build.gradle.kts
@@ -69,10 +69,4 @@ publishing {
             }
         }
     }
-    repositories {
-        maven {
-            name = "test"
-            url = uri("${rootProject.buildDir}/repos/test")
-        }
-    }
 }
